refactor: 패키지 구조 수정 및 코드 리팩토링
