+++ src/main/java/com/todoay/api/domain/auth/controller/AuthController.java
@@ -4,7 +4,7 @@ import com.todoay.api.domain.auth.dto.AuthSaveDto;
 import com.todoay.api.domain.auth.repository.AuthRepository;
 
 import com.todoay.api.domain.auth.service.AuthServiceImpl;
-import com.todoay.api.domain.profile.service.ProfileService;
+//import com.todoay.api.domain.profile.service.ProfileService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -18,7 +18,7 @@ import org.springframework.web.bind.annotation.RestController;
 public class AuthController {
 
     private final AuthServiceImpl authServiceImpl;
-    private final ProfileService profileService;
+//    private final ProfileService profileService;
 
     @PostMapping("save")
     public AuthSaveDto signup(@RequestBody AuthSaveDto authSaveDto) {
