+++ src/main/java/com/todoay/api/domain/auth/service/MailVerificationService.java
@@ -3,7 +3,7 @@ package com.todoay.api.domain.auth.service;
 import com.todoay.api.domain.auth.dto.EmailDto;
 import com.todoay.api.domain.auth.dto.EmailTokenDto;
 import com.todoay.api.domain.auth.utility.MailHandler;
-import com.todoay.api.global.jwt.JwtManager;
+import com.todoay.api.global.security.jwt.JwtManager;
 import io.jsonwebtoken.Claims;
 import lombok.RequiredArgsConstructor;
 import org.springframework.mail.javamail.JavaMailSender;
