+++ src/main/java/com/todoay/api/global/security/jwt/JwtAuthenticationEntryPoint.java
@@ -1,4 +1,4 @@
-package com.todoay.api.global.jwt;
+package com.todoay.api.global.security.jwt;
 
 import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.web.AuthenticationEntryPoint;
@@ -8,6 +8,7 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
+// API 요청에 대해 토큰 인증을 실패했을 때에 대한 처리
 @Component
 public class JwtAuthenticationEntryPoint implements AuthenticationEntryPoint {
