+++ src/main/java/com/todoay/api/global/security/jwt/JwtDto.java
@@ -1,10 +1,10 @@
-package com.todoay.api.global.jwt;
+package com.todoay.api.global.security.jwt;
 
 import lombok.Builder;
 import lombok.Data;
 
 @Data @Builder
-public class TokenDto {
+public class JwtDto {
     private String grantType;
     private String accessToken;
     private String refreshToken;
